feat: Refactor to use new permissions model
The new permissions model should make it possible to avoid reindexing
when a domain changes. A user's permissions determine which filters need
to be applied when searching, so a significant part of this commit is
the refactor of the search logic.

Tests have been changed to use a mock for Elasticsearch to allow request
expectations to be tested.